+++ app/src/main/java/hu/bme/aut/tvshows/ui/episodedetail/EpisodeDetailContract.kt
@@ -9,7 +9,7 @@ interface EpisodeDetailContract {
     }
 
     interface Presenter {
-        fun fetchEpisodeDetails(episodeId: Int)
+        fun fetchEpisodeDetails(episodeId: Long)
         fun cleanup()
     }
 }
