+++ app/src/main/java/hu/bme/aut/tvshows/ui/tvshowdetail/TvShowDetailContract.kt
@@ -11,7 +11,7 @@ interface TvShowDetailContract {
     }
 
     interface Presenter {
-        fun getDetails(id: Int)
+        fun getDetails(id: Long)
         fun cleanup()
     }
 }
